+++ cf-utils/cf-cli/README.md
@@ -6,7 +6,7 @@ It applies the arguments and configure connectors and endpoints.
 
 This module supports coap (plain UDP) and coaps (over DTLS).
 
-For TCP support, a second module "cf-cli-netty-tcp" is provided. If TCP is required, use it by copying the built library of that "cf-cli-netty-tcp" into the folder of the client's jar, either manually or by maven plugin in the client's pom:
+For TCP support, a second module "cf-cli-tcp-netty" is provided. If TCP is required, use it by copying the built library of that "cf-cli-tcp-netty" into the folder of the client's jar, either manually or by maven plugin in the client's pom:
 
 ```xml
 	<plugin>
@@ -22,7 +22,7 @@ For TCP support, a second module "cf-cli-netty-tcp" is provided. If TCP is requi
 					<artifactItems>
 						<artifactItem>
 							<groupId>${project.groupId}</groupId>
-							<artifactId>cf-cli-netty-tcp</artifactId>
+							<artifactId>cf-cli-tcp-netty</artifactId>
 							<version>${project.version}</version>
 							<type>${project.packaging}</type>
 						</artifactItem>
@@ -34,16 +34,33 @@ For TCP support, a second module "cf-cli-netty-tcp" is provided. If TCP is requi
 	</plugin>
 ```
 
-Alternatively adapt the dependency from "cf-cli" to "cf-cli-netty-tcp" in the client's pom.
+If the application is intended to be start by "java -jar xxx.jar", also a classpath entry in the manifest is required.
+
+```xml
+	<plugin>
+		<artifactId>maven-assembly-plugin</artifactId>
+		<configuration>
+			<archive>
+				<manifestEntries>
+					<!-- support tcp, if module library is available -->
+					<Class-Path>cf-cli-tcp-netty-${project.version}.jar</Class-Path>
+				</manifestEntries>
+			</archive>
+		</configuration>
+	</plugin>
+```
+
+Alternatively adapt the dependency from "cf-cli" to "cf-cli-tcp-netty" in the client's pom.
 
 ```xml
   <dependencies>
     ...
     <dependency>
             <groupId>org.eclipse.californium</groupId>
-            <artifactId>cf-cli-netty-tcp</artifactId>
+            <artifactId>cf-cli-tcp-netty</artifactId>
             <version>${project.version}</version>
     </dependency>
     ...
   </dependencies>
 ```
+
