+++ demo-apps/cf-extplugtest-client/pom.xml
@@ -48,7 +48,7 @@
 					<archive>
 						<manifestEntries>
 							<!-- support tcp, if module library is available -->
-							<Class-Path>cf-cli-netty-tcp-${project.version}.jar</Class-Path>
+							<Class-Path>cf-cli-tcp-netty-${project.version}.jar</Class-Path>
 						</manifestEntries>
 					</archive>
 				</configuration>
@@ -66,7 +66,7 @@
 							<artifactItems>
 								<artifactItem>
 									<groupId>${project.groupId}</groupId>
-									<artifactId>cf-cli-netty-tcp</artifactId>
+									<artifactId>cf-cli-tcp-netty</artifactId>
 									<version>${project.version}</version>
 									<type>${project.packaging}</type>
 								</artifactItem>
